# backend/docs/openapi.yaml
openapi: 3.0.3
info:
  title: 1Vault Backend API
  version: 1.2.2
  description: |
    REST API for 1Vault (Solana vaults). Pass `cluster=devnet|mainnet-beta` or header `X-1Vault-Cluster`.

    **Frontend integration guide:** `backend/docs/FRONTEND_GUIDE.md` (vault types, 1vault Licence / 1VL,
    withdraw account layout, flow modes, simulator-v2 canvas mapping).

    **Groups**
    - **Token · Market** — spot price / kline chart
    - **Token · Research** — token info / security / pool / holders / traders / holder analysis
    - **Wallet · Analytics** — portfolio, activity, stats, created tokens, copy-trade score
    - **Discover** — trending profiles, boosts, metas, pair search
    - **Live · Stream** — WebSocket live feeds (profiles / boosts / takeovers / ads)
    - **Account · Wallets** — bind Solana wallet to logged-in user
    - **Tx · *** — unsigned tx prep (client signs → `/v1/tx/submit`)
    - **Flows** — multi-step async jobs

    Token market, research, and wallet analytics require market data to be configured.
    Token metadata (`name` / `symbol` / links) is untrusted user-facing data.

    **Wallet kind:** `walletKind=eoa` (individual/personal) vs `walletKind=pda` (vault wallet).
    See Wallet · Analytics tag and `GET /v1/wallets/{walletAddress}/kind`.

    **Roles:** `strategies` (formerly degen) and `investors` (formerly retail). Legacy `degen`/`retail` still accepted on write.

...
servers:
  - url: http://localhost:3090
    description: Local
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    ClusterQuery:
      name: cluster
      in: query
      required: true
      description: |
        Vault program cluster (`devnet` or `mainnet-beta`). Selects on-chain program addresses.
        Token/wallet market analytics still use Solana mainnet market data regardless of this value.
      schema:
        type: string
        enum: [devnet, mainnet-beta]
...
    VaultTypeQuery:
      name: vaultType
      in: query
      required: false
      schema:
        type: string
        enum: [pooled, sliced]
      description: |
        Product vault type filter.
        - `pooled` — Pooled Vault (single pool / shared strategy)
        - `sliced` — Sliced Vault (capital split across slices)
...
  schemas:
...
    RolePreference:
      type: string
      enum: [strategies, investors]
      description: |
        Product role language:
        - `strategies` — Strategies side (formerly degen)
        - `investors` — Investors side (formerly retail)
        Legacy values `degen` / `retail` are still accepted on write and mapped automatically.
...
    LedgerDepositCreateBody:
      type: object
      required: [vault, investor, role, amount]
      properties:
        vault: { type: string }
        investor:
          type: string
          description: Must be a bound EOA wallet of the caller
        role:
          $ref: "#/components/schemas/RolePreference"
        amount:
          type: string
          description: Lamports as decimal digit string
        takeProfitBps: { type: integer }
        stopLossBps: { type: integer }
    LedgerDepositSubmitBody:
      type: object
      required: [signature]
      properties:
        signature:
          type: string
          description: On-chain tx signature length at least 32
    LedgerDepositFailBody:
      type: object
      properties:
        error:
          type: string
          default: failed
          description: Truncated to 500 chars
    LedgerMandateBody:
      type: object
      required: [vault, investor]
      properties:
        vault: { type: string }
        investor: { type: string }
        role:
          $ref: "#/components/schemas/RolePreference"
        parkAmount:
          type: string
          default: "0"
        takeProfitBps: { type: integer }
        stopLossBps: { type: integer }
        autoFollow:
          type: boolean
          default: true
    IngestBody:
      type: object
      required: [signature]
      properties:
        signature: { type: string }
...
paths:
...
  /v1/vaults:
    get:
      tags: [Vaults]
      summary: List indexed vaults
      description: |
        Paginated list of vaults known to the indexer for the cluster. Entry point for vault discovery/browse UI.
        Each item includes `vaultType` / `vaultTypeLabel` / `vaultTypeMeaning` (default `pooled`).
      parameters:
        - $ref: "#/components/parameters/ClusterQuery"
        - $ref: "#/components/parameters/VaultTypeQuery"
      responses:
        "200":
          description: Paginated vaults
  /v1/vaults/{pubkey}:
    get:
      tags: [Vaults]
      summary: Get vault detail
      description: |
        Single vault record (config, strategist, fees, status, etc.). The vault pubkey is a **PDA** wallet
        (`walletKind=pda`, `walletKindLabel=vault`). Also returns `vaultType*` fields. Use for vault detail pages.
      parameters:
        - $ref: "#/components/parameters/ClusterQuery"
        - name: pubkey
          in: path
          required: true
          schema: { type: string }
          description: On-chain account pubkey
      responses:
        "200":
          description: Vault detail
  /v1/vaults/{pubkey}/holdings:
    get:
      tags: [Vaults]
      summary: Vault token holdings
      description: |
        Holder book for a vault. Response includes `walletKind=pda` (vault wallet) and `vaultType*` fields.
      parameters:
        - $ref: "#/components/parameters/ClusterQuery"
        - name: pubkey
          in: path
          required: true
          schema: { type: string }
          description: On-chain account pubkey
      responses:
        "200":
          description: Holdings
  /v1/vaults/{pubkey}/positions:
    get:
      tags: [Vaults]
      summary: Vault open positions
      description: |
        Open trading positions tracked for this vault.
      parameters:
        - $ref: "#/components/parameters/ClusterQuery"
        - name: pubkey
          in: path
          required: true
          schema: { type: string }
          description: On-chain account pubkey
      responses:
        "200":
          description: Positions
  /v1/vaults/{pubkey}/fees:
    get:
      tags: [Vaults]
      summary: Vault fee accounting
      description: |
        Fee-related state/history for the vault (accrued/claimed performance fees, etc.).
      parameters:
        - $ref: "#/components/parameters/ClusterQuery"
        - name: pubkey
          in: path
          required: true
          schema: { type: string }
          description: On-chain account pubkey
      responses:
        "200":
          description: Fees
  /v1/vaults/{pubkey}/trades:
    get:
      tags: [Vaults]
      summary: Vault trade history
      description: |
        Indexed trades executed by this vault.
      parameters:
        - $ref: "#/components/parameters/ClusterQuery"
        - name: pubkey
          in: path
          required: true
          schema: { type: string }
          description: On-chain account pubkey
      responses:
        "200":
          description: Trades
...
  /v1/vaults/{pubkey}/payouts:
    get:
      tags: [Vaults]
      summary: Vault close payouts
      description: |
        Payout records related to vault close / investor distributions.
      parameters:
        - $ref: "#/components/parameters/ClusterQuery"
        - name: pubkey
          in: path
          required: true
          schema: { type: string }
          description: On-chain account pubkey
      responses:
        "200":
          description: Close payouts
  /v1/vaults/{pubkey}/follows:
    get:
      tags: [Vaults]
      summary: Vault follow events
      description: |
        Investor follow on/off history for this vault.
      parameters:
        - $ref: "#/components/parameters/ClusterQuery"
        - name: pubkey
          in: path
          required: true
          schema: { type: string }
          description: On-chain account pubkey
      responses:
        "200":
          description: Follow events
  /v1/leaderboard:
    get:
      tags: [Actors]
      summary: Top 10 vault leaderboard
      description: |
        Returns at most **10** vaults ranked by profit (`return_pct`) descending.
        Tie-break: higher `nav` first.
      parameters:
        - $ref: "#/components/parameters/ClusterQuery"
      responses:
        "200":
          description: "`{ items (≤10), limit: 10, orderBy: return_pct }`"
  /v1/trades:
    get:
      tags: [Actors]
      summary: List indexed trades
      description: |
        Trade feed from the indexer. Omit `vault` for a global recent feed, or pass `vault` to filter one vault PDA.
      parameters:
        - $ref: "#/components/parameters/ClusterQuery"
        - name: vault
          in: query
          required: false
          schema: { type: string }
          description: Optional vault pubkey filter
        - name: limit
          in: query
          required: false
          schema: { type: integer, default: 100, minimum: 1, maximum: 200 }
      responses:
        "200":
          description: "`{ items, vault, limit }`"
...
  /v1/ledger/deposits:
    get:
      tags: [Ledger]
      security: [{ bearerAuth: [] }]
      summary: List deposit intents
      description: |
        Lists off-chain deposit intents for an investor (optional vault filter).
      parameters:
        - $ref: "#/components/parameters/ClusterQuery"
        - name: investor
          in: query
          required: true
          schema: { type: string }
        - name: vault
          in: query
          required: false
          schema: { type: string }
      responses:
        "200":
          description: "`{ items: DepositIntent[] }` (roles rewritten to strategies|investors)"
    post:
      tags: [Ledger]
      security: [{ bearerAuth: [] }]
      summary: Create deposit intent
      description: |
        Creates an off-chain deposit intent before (or alongside) preparing an on-chain park tx.
        Investor must be a bound wallet of the authenticated user.
      parameters:
        - $ref: "#/components/parameters/ClusterQuery"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LedgerDepositCreateBody"
      responses:
        "201":
          description: "`{ id, status: pending, role, roleLabel }`"
  /v1/ledger/deposits/{id}/submit:
    post:
      tags: [Ledger]
      security: [{ bearerAuth: [] }]
      summary: Mark deposit intent submitted
      description: |
        Marks a deposit intent as submitted after the on-chain tx is broadcast.
      parameters:
        - $ref: "#/components/parameters/ClusterQuery"
        - name: id
          in: path
          required: true
          schema: { type: integer }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LedgerDepositSubmitBody"
      responses:
        "200":
          description: "`{ id, signature, status: submitted }`"
  /v1/ledger/deposits/{id}/fail:
    post:
      tags: [Ledger]
      security: [{ bearerAuth: [] }]
      summary: Mark deposit intent failed
      description: |
        Marks a deposit intent as failed (user cancelled / tx failed).
      parameters:
        - $ref: "#/components/parameters/ClusterQuery"
        - name: id
          in: path
          required: true
          schema: { type: integer }
      requestBody:
        required: false
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LedgerDepositFailBody"
      responses:
        "200":
          description: "`{ id, status: failed }`"
  /v1/ledger/mandates:
    get:
      tags: [Ledger]
      security: [{ bearerAuth: [] }]
      summary: List investor mandates
      description: |
        Lists investor mandate/config records stored off-chain.
      parameters:
        - $ref: "#/components/parameters/ClusterQuery"
        - name: investor
          in: query
          required: true
          schema: { type: string }
        - name: vault
          in: query
          required: false
          schema: { type: string }
      responses:
        "200":
          description: "`{ items: Mandate[] }`"
    post:
      tags: [Ledger]
      security: [{ bearerAuth: [] }]
      summary: Upsert investor mandate
      description: |
        Create or update an investor mandate (risk limits / follow preferences stored off-chain).
        Defaults: `role=investors`, `parkAmount="0"`, `autoFollow=true`.
      parameters:
        - $ref: "#/components/parameters/ClusterQuery"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LedgerMandateBody"
      responses:
        "200":
          description: "`{ ok: true, role, roleLabel }`"
  /v1/ingest:
    post:
      tags: [System]
      summary: Admin indexer ingest hook
      description: |
        Admin-only endpoint to push a signature to the indexer. Requires `X-Admin-Key`.
      parameters:
        - $ref: "#/components/parameters/ClusterQuery"
        - name: X-Admin-Key
          in: header
          required: true
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/IngestBody"
      responses:
        "200":
          description: Proxied indexer response
...